relocate Source folder
When going back to an older commit corresponding with version 1.5.3.
The Source folder containing the template information and input parameters was out of the scope of git and updated but not with backwards compatibility. Therefore it has been relocated to the SpotNGlia folder and an older version of Template3dpf (struct )from the 1.5.2 build replaced the newer version (object. In SpotnGlia.m the SourcePath location is changed.