renaming and relocating used functions
This commit is different from the equal named commit in the master branch. Some functions in the develop branch are already evolved but this is not visible in the master branch due to a strategy=ours merge. commits between ~feb/mrt 2018 to jul 19, 2018 are not present in latest master build.

-single funtions in 'SpotNGlia' folder are placed in folder 'main functions'
-SpotOptimization is removed to SpotNGlia-private
-'sng_functions' is renamed to 'support functions'